Restructure and footer changes
