+++ fitznet-react-website/src/App.js
@@ -1,11 +1,12 @@
-import './App.css';
+import './css/App.css';
 import React from 'react';
 import { BrowserRouter, Outlet, Route, Routes } from 'react-router-dom';
-import Homepage from './Homepage';
-import NoPage from './NoPage';
-import Navbar from './Navbar';
-import LaurenPanelContent from './LaurenPanelContent';
-import InfoPanelContent from './InfoPanelContent';
+import Homepage from './components/Homepage';
+import NoPage from './components/NoPage';
+import Navbar from './components/Navbar';
+import LaurenPanelContent from './components/LaurenPanelContent';
+import InfoPanelContent from './components/InfoPanelContent';
+import Footer from './components/Footer';
 
 
 function App() {
@@ -19,6 +20,7 @@ function App() {
         <Route path="/info" element={<InfoPanelContent />} />
         <Route path="*" element={<NoPage />} />
       </Routes>
+      <Footer/>
       <Outlet />
     </BrowserRouter>
   );
