+++ fitznet-react-website/src/components/InfoPanelContent.js
