+++ fitznet-react-website/src/components/LaurenPanelContent.js
