+++ fitznet-react-website/src/components/Navbar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './Navbar.css';
+//import './css/Navbar.css  ';
 
 function Navbar() {
   return (
