+++ fitznet-react-website/src/css/Footer.css
@@ -1,0 +1,32 @@
+.footer {
+    background-color: #333;
+    color: #fff;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem;
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+    font-size: 0.8rem;
+    text-align: center;
+  }
+  
+  .footer a {
+    color: #fff;
+  }
+  
+  @media screen and (min-width: 768px) {
+    .footer {
+      flex-direction: row;
+      justify-content: space-between;
+      font-size: 1rem;
+      padding: 1rem 2rem;
+    }
+    
+    .footer-text {
+      margin-right: 2rem;
+    }
+  }
+
